+++ src/common/game.d.ts
@@ -67,7 +67,9 @@ export interface GameState {
 export interface GameScore {
   cells: number,
   mines: number,
+  level?: number,
   playMode?: PlayMode,
+  mode?: PlayMode,
   effort: {
     least: number,
     most: number,
@@ -97,7 +99,6 @@ export interface BareScoreItem {
 
 export interface ScoreItem extends BareScoreItem {
   rank: number,
-  level: number,
 }
 
 export type ScoreParam = keyof ScoreItem | keyof ScoreItem["game"] | keyof ScoreItem["play"] | keyof ScoreItem["score"]